Simplify construction of implicit providers
