+++ form/form_source.cpp
@@ -112,12 +112,12 @@ namespace {
           return this->read_product_from_form(actual_creator_, name, id.to_string(), product_type);
         };
 
-        bundles.push_back(phlex::detail::provider_bundle{
-          .provider_function = std::function<phlex::detail::provider_function_t>(provider_func),
-          .max_concurrency = phlex::concurrency::serial,
-          .spec = std::move(spec),
-          .layer = std::string(selector_layer.trans_get_string()),
-          .stage = std::string(selector_stage.trans_get_string())});
+        bundles.push_back(
+          phlex::detail::provider_bundle{.provider_function = provider_func,
+                                         .max_concurrency = phlex::concurrency::serial,
+                                         .spec = std::move(spec),
+                                         .layer = std::string(selector_layer.trans_get_string()),
+                                         .stage = std::string(selector_stage.trans_get_string())});
       }
 
       return bundles;
